+++ contents/config/config.qml
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: 2026 Thuan Phat <laithuanphat@gmail.com>
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+import QtQuick
+import org.kde.plasma.configuration
+
+ConfigModel {
+    ConfigCategory {
+        name: i18n("General")
+        icon: "preferences-desktop"
+        source: "configGeneral.qml"
+    }
+}